+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "editor.quickSuggestions": {
+        "other": "on",
+        "comments": true,
+        "strings": "off"
+    }
+}
